% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames.R
\name{colname-functions}
\alias{colname-functions}
\alias{colname_set}
\alias{colname_get}
\alias{colname_time}
\alias{colname_time<-}
\alias{colname_value}
\alias{colname_value<-}
\alias{colname_signal}
\alias{colname_signal<-}
\alias{colname_fits}
\alias{colname_fits<-}
\title{Column-name related functions}
\usage{
colname_set(x, value, name)

colname_get(x, name)

colname_time(x)

colname_time(x) <- value

colname_value(x)

colname_value(x) <- value

colname_signal(x)

colname_signal(x) <- value

colname_fits(x)

colname_fits(x) <- value
}
\arguments{
\item{x}{A \code{\link{tbl_mbte}}.}

\item{value}{The name of the column as a \code{\link[base:name]{symbol}}.}

\item{name}{A character vector of length 1 (string) describing the purpose
of the column to get/set (e.g. \code{"signal"} or \code{"fits"}).}
}
\value{
Replacement functions like \code{colname_time<-} return a modified
  \code{\link{tbl_mbte}}. Otherwise a \code{\link[base:name]{symbol}} is
  returned.
}
\description{
Internal helper functions to access colnames used in a specific dataset;
Attributes are currently used to store column-names as symbols (to avoid
having a custom datastructure, since the dataset should still be tidyverse-
friendly)
}
\section{Functions}{
\itemize{
\item \code{colname_set}: Set a specific column name. This function is
designed for extension-packages.

\item \code{colname_get}: Get a specific column name. This function is
designed for extension-packages.

\item \code{colname_time}: Get column name for signal-time.

\item \code{colname_time<-}: Override column name for signal-time.

\item \code{colname_value}: Get column name for signal-values.

\item \code{colname_value<-}: Set column name for signal-values column.

\item \code{colname_signal}: Get column name for signal-column.

\item \code{colname_signal<-}: Override column name for signal-column.

\item \code{colname_fits}: Get column name for fits-column (see
\code{\link{mbte_fit}}).

\item \code{colname_fits<-}: Set column name for fits column.
}}

\seealso{
\code{\link{tbl_mbte}} (explanation of used columns)
}
\keyword{internal}
