% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-store.R, R/attributes.R
\name{event-logging}
\alias{event-logging}
\alias{mbte_event_log}
\title{Log unusual events occurring during execution}
\usage{
mbte_event_log(x)
}
\arguments{
\item{x}{A \code{\link{tbl_mbte}} with event-log information stored in
attributes.}
}
\value{
A tibble with logged event-information. Each row represents an event.
The column layout is specific to the function, which produced the event-log.
}
\description{
For some occasions (like in long-running computations) it may not be desired
to shut down the whole computation if a single error occurrrs. It may be
better to log the occurred error and carry on with the computation. In such
a case, \code{\link{NA}} will always be the result of a failed computation.
The logged information can be retrieved via \code{mbte_event_log()} (see
examples).
}
\examples{
library(dplyr, warn.conflicts = FALSE)
data(raw_signals)
raw_signals

# prepare for signal extraction (perform conversion to `tbl_mbte` and nest
# signals)
raw_signals <- raw_signals \%>\%
  group_by(mv) \%>\%
  new_tbl_mbte(time = "t", value = "value") \%>\%
  mbte_nest_signals()

# provoke an error by passing an indexing-function, that will always raise
# an error
faulty_extract <- mbte_extract_subsignals(raw_signals, f = function(x) {
  stop("test")
})

# retrieve event-log
event_log <- mbte_event_log(faulty_extract)
head(event_log)

# show occurred errors
head(event_log$error)

}
\seealso{
\code{\link{raw_signals}} (dataset in examples)
}
