% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{fitting-helpers}
\alias{fitting-helpers}
\alias{tr_linear}
\alias{tr_exponential}
\alias{tr_logistic}
\title{Trend-fitting helpers}
\usage{
tr_linear()

tr_exponential()

tr_logistic()
}
\value{
A fitting-\code{\link[rlang:quotation]{quosure}} will be returned.
}
\description{
Helper functions meant to be used in conjunction with
\code{\link{mbte_fit}}.
}
\section{Functions}{
\itemize{
\item \code{tr_linear}: Determine a linear trend using
\code{\link[stats]{lm}}.

\item \code{tr_exponential}: The presence of an exponenetial trend is
determined by applying a log-transformation to the signal-values and
fitting them via \code{\link[stats]{lm}}. A vector of the correct length
with \code{NA}'s will be returned if some values of the signal are zero or
negative.

\item \code{tr_logistic}: Fit a sigmoid model using
  \code{\link[nls2]{nls2}}. This fitting helper requires the `nls2`-package
  (the "port"-algorithm with 500 iterations is used). The signal-tibble to
  fit must have at least five rows (see \code{\link{mbte_fit}}). Otherwise
  \code{NA}'s are returned.

  The formula used for fitting is
  \code{value ~ A / (1 + exp(B * (C - time))) + D}. NOTE: warnings are
  suppressed.
}}

\examples{
data(filtered_signals)

# Use fitting helpers to fit signals
#
# NOTE: The unquoting operator (`!!`) is used, since fitting helpers return
# quosures.
fits1 <- mbte_fit(filtered_signals, lin = !!tr_linear())
fits1

# Equivalent without using fitting-helpers (columns for time and value of
# signal have to be provided explicitly)
fits2 <- mbte_fit(filtered_signals, lin = lm(value ~ t, .signal))
fits2

}
\seealso{
\code{\link{filtered_signals}} (dataset used in examples)
}
