% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subsignals.R
\name{indexing-function}
\alias{indexing-function}
\title{Indexing functions for subsignal-extraction}
\value{
A list with the following names should be returned:
\describe{
  \item{start}{Start indices of the found subsignals.}
  \item{end}{End indices of the found subsignals.}
}
}
\description{
Indexing functions have the following signature:
\code{function(x, ...)}. \code{x} denotes a signal-values vector (see
\code{\link{tbl_mbte}} for  details). In other words, the measured values of
a signal get passed as a numeric vector. The job of an indexing function is
to find the start- and end-indices of a subsignal (used by
\code{\link{mbte_extract_subsignals}}).
}
\note{
The length of the returned indices must match the number of found subsignals
(e.g. 3 Subsignals found ==> length of start- and end-indices vectors must
equal 3). The indices-vectors have to be of type integer.
}
