% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_metrics.R
\name{mbte_compute_metrics}
\alias{mbte_compute_metrics}
\title{Compute metrics based on the fitted signals and the original signal-values.}
\usage{
mbte_compute_metrics(x, ...)
}
\arguments{
\item{x}{A `tbl_mbte`, where predicted signals and original signals are
present.}

\item{...}{The ellipsis must only contain named elements. The elements are
used as quosures and tidy evaluation is used. Caution:
`.pred` and `.obs` are masked (See details for more information.).
Passed closures must evaluate to a scalar numeric (double). Otherwise or if
an error is encountered, \code{NA_real_} will be returned.}
}
\description{
Compute metrics based on the fitted signals and the original signal-values.
}
\details{
The following objects are masked via a data-mask:
\describe{
  \item{.pred}{The predicted signal-values}
  \item{.obs}{The observed (original) signal-values}
}
}
\section{event-table}{

The error-log (tibble) contains the following columns:
\describe{
  \item{error}{The error, which occurred during processing. Errors occurring
    during the evaluation of a metric-quosure (a element passed to
    \code{...}) are wrapped.}
  \item{row_nr}{The row-number of the original table (\code{x}), at which the
    error occurred.}
  \item{fit_name}{The name of the fit.}
  \item{metric_name}{The name of the current metric (a name in \code{...}).}
  \item{metric_quo}{The current metric-quosure (element of \code{...}).}
  \item{pred}{The current predicted signal-values (expected to be be
    numeric).}
  \item{obs}{The observed signal-values (of the original signal).}
}
}

\section{event-logging}{

This function logs unusual events. A warning gets raised at the end of
execution, if events have been logged. The error log can be retrieved
by passing the returned object to \code{\link{mbte_event_log}}.
}

