% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subsignals.R
\name{mbte_default_indexer}
\alias{mbte_default_indexer}
\title{The default indexer for the mbte-package}
\usage{
mbte_default_indexer(x, ...)
}
\arguments{
\item{x}{A numeric vector}

\item{...}{Additional arguments (currently ignored)}
}
\description{
The default indexer for the mbte-package
}
