% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.R
\name{mbte_event_log}
\alias{mbte_event_log}
\title{Retrieve logged information about occurred unusual events}
\usage{
mbte_event_log(x)
}
\arguments{
\item{x}{A `tbl_mbte` with event-log information stored in attributes.}
}
\value{
A tibble with logged event-information. Each row represents an event.
}
\description{
Retrieve logged information about occurred unusual events
}
