% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_subsignals.R
\name{mbte_extract_subsignals}
\alias{mbte_extract_subsignals}
\title{Extract subsignals from a `tbl_mbte`-object}
\usage{
mbte_extract_subsignals(x, f = mbte_default_indexer, ...)
}
\arguments{
\item{x}{A `tbl_mbte`-object}

\item{f}{An indexing function (a function taking a numeric vector and
returning a list). The returned list should contain the elements `start` and
`end` (interger-vector of the starting- or ending-positions of the
subsignal respecitvely).}

\item{...}{Additional arguments passed to `f`}
}
\value{
The original table gets returned. The "signal"-column
is modified (since subsignals are extracted according to the indexing
function \code{f}).
}
\description{
Extract subsignals from a `tbl_mbte`-object
}
\section{event-logging}{

This function logs unusual events. A warning gets raised at the end of
execution, if events have been logged. The error log can be retrieved
by passing the returned object to \code{\link{mbte_event_log}}.
}

\section{event-table}{

A tibble containing event-information consists of 3 columns:
\describe{
  \item{error}{The unprocessed error, which occurred during execution.}
  \item{row_nr}{The row-number of the input-tibble (\code{x}), at which the
    error occurred.}
  \item{signal}{The signal-subtable processed at the time the error occurred.}
}
}

