% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mbte_panel_plot}
\alias{mbte_panel_plot}
\title{Create a basic plot of signals or fits}
\usage{
mbte_panel_plot(x, expr, ..., nrow = 3, ncol = 4)
}
\arguments{
\item{x}{A \code{\link{tbl_mbte}}.}

\item{expr}{An expression to generate one panel of the plot (gets quoted)}

\item{...}{Variables used for grouping (get quoted).}

\item{nrow}{Number of rows per page.}

\item{ncol}{Number of columns per page.}
}
\value{
A list of ggplot-objects.
}
\description{
A \code{\link{tbl_mbte}} is passed. It is assumed, that the table is ordered
(panels to draw first are at the top of the table). The contents of the
panels (subplots) to draw are determined by the grouping specified in the
ellipsis (\code{...}). Since the number of panels to draw may not fit onto
one plot (can only contain \code{nrow * ncol} elements), panels are split
onto mulitple "pages". This is done by returning a list, where each element
is a ggplot2-chart, showing a subset of the data to plot.
}
\details{
The user should pass a \code{\link[rlang:quotation]{quosure}} as \code{expr},
which specifies what should be drawn in a panel/subplot. It can make use of
the following masked objects:

\describe{
  \item{.u_signals}{Unnested signals (see \code{\link{mbte_unnest_signals}})}
  \item{.u_fits}{Unnested fits (see \code{\link{mbte_unnest_fits}})}
  \item{.time_sym}{Time variable as a \code{\link[base:name]{symbol}}}
  \item{.value_sym}{Value-variable as a \code{\link[base:name]{symbol}}}
  \item{.n_pages}{The total number of pages, that will be printed}
  \item{.n_page}{The number of the current page (when evaluating
    \code{expr})}
}
}
\note{
\code{.u_signals} and \code{.u_fits} are only available if the `signal`
and `fits`-columns are present.
}
\examples{
library(ggplot2)

# load dataset
data(filtered_signals)
filtered_signals

# Draw a graph grouped by the measurement variable (multiple subsignals will
# be displayed in one panel/subplot).
plots <- mbte_panel_plot(
  x = filtered_signals,
  expr = {
    # expression for the content a subplot/panel
    # NOTE: `.u_signals` is provided by masking
    ggplot(.u_signals, aes(t, value)) +
      geom_path(aes(group = signal_nr))
  },
  mv # grouping varibales
)

# only show 2 "pages"
head(plots, n = 2)

# Now grouping by measurement variable (measured parameter) and signal_nr is
# performed.
plots <- mbte_panel_plot(
  x = filtered_signals,
  expr = {
    # NOTE: Using a grouping aesthetic is not necessary, since only one line
    # per panel will be drawn (because a signal in the used dataset is
    # identified by the `mv` and the `signal_nr` column).
    ggplot(.u_signals, aes(t, value)) +
      geom_path()
  },
  mv, signal_nr # grouping variables
)

# only display 2 elements of returned list
head(plots, n = 2)

}
\seealso{
\code{\link{filtered_signals}} (dataset used in examples),
\code{\link[ggplot2]{facet_wrap}}
}
