% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{mbte_reconstruct}
\alias{mbte_reconstruct}
\alias{mbte_reconstruct.tbl_mbte}
\title{Reconstruct S3-tbl_mbte-class from object}
\usage{
mbte_reconstruct(new, old)

\method{mbte_reconstruct}{tbl_mbte}(new, old)
}
\arguments{
\item{new}{new object}

\item{old}{Object, from which specific attributes are copied (time, value,
signal, fits). The S3-generic dispatches on this argument.

The implementation of `sloop::reconstruct()` gets used; This is a temporary
workaround and will be removed, once the sloop-package is on CRAN.}
}
\description{
This function should be used to readd dropped-attributes, if a
`tbl_mbte`-object gets processed by functions, which may drop additional
attributes. The mbte-package relies on attributes in order to store
information about column-names (e.g. name of the time-column) via attributes.
}
\seealso{
Other tbl_mbte functions: \code{\link{is_tbl_mbte}},
  \code{\link{new_tbl_mbte}}, \code{\link{tbl_mbte}}
}
\concept{tbl_mbte functions}
