% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{raw_signals}
\alias{raw_signals}
\title{Sample dataset of raw-signals}
\description{
A \code{\link[tibble]{tibble}} with randomly generated and purely fictional
signal-data. The intensities of fictional measurement variables are measured
every minute. The dataset contains 100 measurements of 42 measurement
variables. A measurement variable can be thought of a variable describing the
measured parameter.
}
\section{Contained colums}{

\itemize{
  \item t. The time column.
  \item mv. The measurement variable.
  \item value. The measured intensity of the signal.
}
}

\seealso{
Other datasets: \code{\link{filtered_signals}}
}
\concept{datasets}
\keyword{data}
