% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_mbte.R
\name{is_tbl_mbte}
\alias{is_tbl_mbte}
\title{Checks if an object is a \code{tbl_mbte}.}
\usage{
is_tbl_mbte(x)
}
\arguments{
\item{x}{The object to check}
}
\description{
Checks if an object is a \code{tbl_mbte}.
}
\details{
Currently, a valid \code{tbl_mbte} must be a tibble, which also
inherits from \code{tbl_mbte}. The attributes \code{time}, \code{value},
\code{signal} and \code{fits} have to be present.
}
