% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{mbte_fit}
\alias{mbte_fit}
\title{Fit models to signals}
\usage{
mbte_fit(x, ...)
}
\arguments{
\item{x}{A `tbl_mbte`, where signals are present as subtibbles in a
list-column.}

\item{...}{It is assumed, that the ellipsis only contains named elements. The
elements are used as quosures and tidy evaluation is used. Caution:
`.signal`, `.time_sym` and `.value_sym` are masked (See  details for more
information.).
Closures may return a numeric vector with the length of the original
signal-values or an object with a method for \code{\link[stats]{predict}}.
If the quosure evaluates to a numeric vector, it gets used as is. Otherwise,
\code{predict(returned_value, newdata = .signal)} will be called, in order to
get the predictions.}
}
\description{
Fit models to signals
}
\details{
Quosures can make use of masked objects, that shadow their normal regular
execution environment:
\describe{
  \item{.signal}{A \code{\link[tibble]{tibble}} with the signal to fit
  (time and value are present)}
  \item{.time_sym}{The name of the time column as a
  \code{\link[base]{name}}.}
  \item{.value_sym}{The name of the value-column stored as a
  \code{\link[base]{name}}}
}
}
\section{event-table}{

The error-log table contains the following columns:
\describe{
  \item{error}{The error, that occurred during processing. Errors which
    originate from the evaluation of a fitting quosure (specified in
    \code{...}) or prediction-related errors are wrapped.}
  \item{row_nr}{The row-number of the original table (\code{x}), at which
    the error occurred.}
  \item{signal}{The signal-subtibble processed at the time the error
    occurred.}
  \item{fit_name}{The name of the fitting quosure (name in \code{...}).}
  \item{fit_quo}{The fitting quosure being processed (element of \code{...}).}
}
}

\section{event-logging}{

This function logs unusual events. A warning gets raised at the end of
execution, if events have been logged. The event-log can be retrieved
by passing the returned object to \code{\link{mbte_event_log}}. In this
case, a tibble containing the logged events will be returned to the user.
}

