% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_signals.R
\name{mbte_nest_signals}
\alias{mbte_nest_signals}
\title{Combine time- and variable- column to a list-column (signal).}
\usage{
mbte_nest_signals(x, ...)
}
\arguments{
\item{x}{A \code{tbl_mbte}, the time- and value-columns are still present (
specified when creating the `tbl_mbte`).}

\item{...}{Variables (get quoted), which should be used for grouping. If no
variables are specified, \code{\link{mbte_nest_signals}} assumes, that
\code{x} is already a grouped table.}
}
\description{
Combine time- and variable- column to a list-column (signal).
}
