% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_mbte.R
\name{new_tbl_mbte}
\alias{new_tbl_mbte}
\title{Creates a new dataset, suited for the the signal detection workflow}
\usage{
new_tbl_mbte(x, time, value, ..., signal = "signal", fits = "fits",
  metric = "metric", subclass = NULL)
}
\arguments{
\item{x}{A data.frame or tibble}

\item{time}{A symbol for the time-column (gets quoted).}

\item{value}{The value column (measurement-data); gets quoted}

\item{...}{Additional attributes, that are set (every element must be named)}

\item{signal}{A name for a list column, in which tibbles containing
the signal (time- and value- column combined) get stored; gets quoted}

\item{fits}{List-column-name; The predicted values for the original signal
get stored in this list-column.}

\item{metric}{The name for numeric-column containing the computed
error-metric.}

\item{subclass}{Additional classes to inherit from (character).}
}
\description{
A function intended to be used internally or to extend mbte.
}
